Step one in shared memory simplification. Make each allocated slot of same size. It'll waste few bytes, but will ensure grater consistency
git-svn-id: https://svn.apache.org/repos/asf/tomcat/jk/trunk@1306209 13f79535-47bb-0310-9956-ffa450edef68

+++ native/apache-1.3/mod_jk.c
@@ -257,7 +257,7 @@ static jk_logger_t *main_log = NULL;
 static table *jk_log_fds = NULL;
 static jk_worker_env_t worker_env;
 static char *jk_shm_file = NULL;
-static size_t jk_shm_size = 0;
+static int jk_shm_size = 0;
 static int jk_shm_size_set = 0;
 /*
  * Worker stuff
@@ -1349,7 +1349,7 @@ static const char *jk_set_shm_size(cmd_parms * cmd,
         sz = JK_SHM_DEF_SIZE;
     else
         sz = JK_SHM_ALIGN(sz);
-    jk_shm_size = (size_t)sz;
+    jk_shm_size = sz;
     if (jk_shm_size)
         jk_shm_size_set = 1;
     return NULL;
